server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pydub==0.25.1
ffmpeg-python==0.2.0